+++ js/main.js
@@ -1,5 +1,5 @@
 import { getData } from './api.js';
-import { renderPhotos } from './renderPhotos.js';
+import { renderPhotos } from './photos.js';
 
 getData()
   .then((data) => {
