+++ js/photos.js
@@ -1,23 +1,12 @@
-const template = document.querySelector('#picture').content.querySelector('.picture');
-const fragment = document.createDocumentFragment();
-const photoList = document.querySelector('.pictures');
+import { renderUploadPhoto } from './upload/upload-photo.js';
+import { filterPhotos } from './upload/filter.js';
 
+const photoPreview = document.querySelector('.img-upload__start');
+const filterElement = document.querySelector('.img-filters');
 
-const createPhoto = (src, description, commentsCount, likesCount) => {
-  const item = template.cloneNode(true);
+photoPreview.addEventListener('change', renderUploadPhoto);
 
-  item.querySelector('img').src = src;
-  item.querySelector('img').alt = description;
-  item.querySelector('.picture__comments').textContent = commentsCount;
-  item.querySelector('.picture__likes').textContent = likesCount;
-
-  return item;
-};
-
-export const renderPhotosSet = (dataArray) => {
-
-  for (let i = 0; i < dataArray.length; i++) {
-    fragment.appendChild(createPhoto((dataArray[i].url), dataArray[i].description, dataArray[i].comments.length, dataArray[i].likes));
-  }
-  return photoList.appendChild(fragment);
+export const renderPhotos = (dataArray) => {
+  filterPhotos(dataArray);
+  filterElement.classList.remove('img-filters--inactive');
 };
