+++ js/upload/preview-upload-photo.js
