+++ js/upload/slider-effects.js
